restructure code

+++ app/Http/Controllers/AboutController.php
@@ -13,8 +13,14 @@ class AboutController extends Controller
      */
     public function index()
     {
-        $about_data = About::All();
-        return view('admin.about',['about_data'=>$about_data]);
+        if(isset(Auth::user()->username)){
+            $about_data = About::All();
+            return view('admin.about',['about_data'=>$about_data]);
+        }
+        if(isset(Auth::user()->username)){
+          
+        }
+        return redirect(route('home'));
     }
 
     /**
@@ -30,28 +36,31 @@ class AboutController extends Controller
      */
     public function store(Request $request)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Create" || trim($userPermission->name) == "create"){
-                // dd($request->heading);
-                $request->validate([
-                    'heading'=>'required',
-                    'content'=>'required'
-                ]);
-                // dd('hello');
-                $image = $request->file('image');
-                $imagename = $image->getClientOriginalName();
-                $request->image->move(public_path('image'),$imagename);
-                $about = new About();
-                $about->heading = $request->heading;
-                $about->content = $request->content;
-                $about->image = $imagename;
-                $about->save();
-                return redirect(route('about'));
-            }else{
-                return redirect(route('dashboard'));
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Create" || trim($userPermission->name) == "create"){
+                    // dd($request->heading);
+                    $request->validate([
+                        'heading'=>'required',
+                        'content'=>'required'
+                    ]);
+                    // dd('hello');
+                    $image = $request->file('image');
+                    $imagename = $image->getClientOriginalName();
+                    $request->image->move(public_path('image'),$imagename);
+                    $about = new About();
+                    $about->heading = $request->heading;
+                    $about->content = $request->content;
+                    $about->image = $imagename;
+                    $about->save();
+                    return redirect(route('about'));
+                }else{
+                    return redirect(route('dashboard'));
+                }
             }
         }
+        return redirect(route('home'));
     }
 
     /**
@@ -67,15 +76,19 @@ class AboutController extends Controller
      */
     public function edit(About $about,$id)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
-                $about = About::where('id',$id)->first();
-                return view('admin.edit_about',['about'=>$about]);
-                
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
+                    $about = About::where('id',$id)->first();
+                    return view('admin.edit_about',['about'=>$about]);
+                    
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+        
     }
 
     /**
@@ -83,29 +96,33 @@ class AboutController extends Controller
      */
     public function update(Request $request, About $about,$id)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
-                $request->validate([
-                    'heading'=>'required',
-                    'content'=>'required',
-                ]);
-
-                $about = About::where('id',$id)->first();
-                $about->heading = $request->heading;
-                $about->content = $request->content;
-                if($request->image != null){
-                    $image = $request->file('image');
-                    $imagename = $image->getClientOriginalName();
-                    $request->image->move(public_path('image'),$imagename);
-                    $about->image = $imagename;
-                } 
-                // dd('nice');
-                $about->save();
-                return redirect(route('about'));
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
+                    $request->validate([
+                        'heading'=>'required',
+                        'content'=>'required',
+                    ]);
+    
+                    $about = About::where('id',$id)->first();
+                    $about->heading = $request->heading;
+                    $about->content = $request->content;
+                    if($request->image != null){
+                        $image = $request->file('image');
+                        $imagename = $image->getClientOriginalName();
+                        $request->image->move(public_path('image'),$imagename);
+                        $about->image = $imagename;
+                    } 
+                    // dd('nice');
+                    $about->save();
+                    return redirect(route('about'));
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+        
     }
 
     /**
@@ -113,14 +130,18 @@ class AboutController extends Controller
      */
     public function destroy(About $about,$id)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Delete" || trim($userPermission->name) == "delete"){
-                $about = About::where('id',$id)->first();
-                $about->delete();
-                return redirect(route('about'));
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Delete" || trim($userPermission->name) == "delete"){
+                    $about = About::where('id',$id)->first();
+                    $about->delete();
+                    return redirect(route('about'));
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+        
     }
 }
