+++ app/Http/Controllers/PrivacyController.php
@@ -13,8 +13,11 @@ class PrivacyController extends Controller
      */
     public function index()
     {
-        $privacys = Privacy::All();
-        return view('admin.privacy_policy',['privacys'=>$privacys]);
+        if(isset(Auth::user()->username)){
+            $privacys = Privacy::All();
+            return view('admin.privacy_policy',['privacys'=>$privacys]);
+        }
+        return redirect(route('home'));
     }
 
     /**
@@ -30,25 +33,29 @@ class PrivacyController extends Controller
      */
     public function store(Request $request)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Create" || trim($userPermission->name) == "create"){
-                $request->validate([
-                    'heading'=>'required',
-                    'content'=>'required'
-                ]);
-                $image = $request->file('image');
-                $imagename = $image->getClientOriginalName();
-                $request->image->move(public_path('image'),$imagename);
-                $privacy = new Privacy();
-                $privacy->heading = $request->heading;
-                $privacy->content = $request->content;
-                $privacy->image = $imagename;
-                $privacy->save();
-                return redirect(route('privecy'));
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Create" || trim($userPermission->name) == "create"){
+                    $request->validate([
+                        'heading'=>'required',
+                        'content'=>'required'
+                    ]);
+                    $image = $request->file('image');
+                    $imagename = $image->getClientOriginalName();
+                    $request->image->move(public_path('image'),$imagename);
+                    $privacy = new Privacy();
+                    $privacy->heading = $request->heading;
+                    $privacy->content = $request->content;
+                    $privacy->image = $imagename;
+                    $privacy->save();
+                    return redirect(route('privecy'));
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+       
     }
 
     /**
@@ -64,15 +71,19 @@ class PrivacyController extends Controller
      */
     public function edit(Privacy $privacy,$id)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
-                $privacy = Privacy::where('id',$id)->first();
-                return view('admin.edit_privacy',['privacy'=>$privacy]);
-                
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
+                    $privacy = Privacy::where('id',$id)->first();
+                    return view('admin.edit_privacy',['privacy'=>$privacy]);
+                    
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+        
     }
 
     /**
@@ -80,30 +91,33 @@ class PrivacyController extends Controller
      */
     public function update(Request $request, Privacy $privacy ,$id)
     {
-       
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
-                $request->validate([
-                    'heading'=>'required',
-                    'content'=>'required',
-                ]);
-
-                $privacy = Privacy::where('id',$id)->first();
-                $privacy->heading = $request->heading;
-                $privacy->content = $request->content;
-                if($request->image != null){
-                    $image = $request->file('image');
-                    $imagename = $image->getClientOriginalName();
-                    $request->image->move(public_path('image'),$imagename);
-                    $privacy->image = $imagename;
-                } 
-                // dd('nice');
-                $privacy->save();
-                return redirect(route('privecy'));
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
+                    $request->validate([
+                        'heading'=>'required',
+                        'content'=>'required',
+                    ]);
+    
+                    $privacy = Privacy::where('id',$id)->first();
+                    $privacy->heading = $request->heading;
+                    $privacy->content = $request->content;
+                    if($request->image != null){
+                        $image = $request->file('image');
+                        $imagename = $image->getClientOriginalName();
+                        $request->image->move(public_path('image'),$imagename);
+                        $privacy->image = $imagename;
+                    } 
+                    // dd('nice');
+                    $privacy->save();
+                    return redirect(route('privecy'));
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+        
     }
 
     /**
@@ -111,14 +125,18 @@ class PrivacyController extends Controller
      */
     public function destroy(Privacy $privacy,$id)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Delete" || trim($userPermission->name) == "delete"){
-                $privacy = Privacy::where('id',$id)->first();
-                $privacy->delete();
-                return redirect(route('privecy'));
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Delete" || trim($userPermission->name) == "delete"){
+                    $privacy = Privacy::where('id',$id)->first();
+                    $privacy->delete();
+                    return redirect(route('privecy'));
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+        
     }
 }
