+++ app/Http/Controllers/RolesController.php
@@ -14,18 +14,19 @@ class RolesController extends Controller
      */
     public function index()
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
+                   
+                $roles = Role::All();
+                return view('admin.user_role',['roles'=>$roles]);
                
-            $roles = Role::All();
-            return view('admin.user_role',['roles'=>$roles]);
-           
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
-            
-          
+        return redirect(route('home')); 
     }
 
     /**
@@ -33,14 +34,17 @@ class RolesController extends Controller
      */
     public function create()
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
-                return view('admin.add_user_role');
-           
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
+                    return view('admin.add_user_role');
+               
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
       
     }
     /**
@@ -48,34 +52,34 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
-                $request->validate([
-                    'role_name' => 'required|max:255',
-                    'role_slug' => 'required|max:255'
-                ]);
-                $role = new Role();
-                $role->name = $request->role_name;
-                $role->slug = $request->role_slug;
-                $role-> save();
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
+                    $request->validate([
+                        'role_name' => 'required|max:255',
+                    ]);
+                    $role = new Role();
+                    $role->name = $request->role_name;
+                    $role-> save();
+                   
+                    $listOfPermissions = explode(',', $request->roles_permissions);//create array from separated/coma permissions
+                    
+                    foreach ($listOfPermissions as $permission) {
+                        $permissions = new Permission();
+                        $permissions->name = $permission;
+                        // $permissions->slug = strtolower(str_replace(" ", "-", $permission));
+                        $permissions->save();
+                        $role->permissions()->attach($permissions->id);
+                        $role->save();
+                    }    
+                    return redirect(route('userrole'));
                
-                $listOfPermissions = explode(',', $request->roles_permissions);//create array from separated/coma permissions
-                
-                foreach ($listOfPermissions as $permission) {
-                    $permissions = new Permission();
-                    $permissions->name = $permission;
-                    $permissions->slug = strtolower(str_replace(" ", "-", $permission));
-                    $permissions->save();
-                    $role->permissions()->attach($permissions->id);
-                    $role->save();
-                }    
-                return redirect(route('userrole'));
-           
-            }else{
-                return redirect(route('dashboard'));
+                }
             }
+            return redirect(route('dashboard'));
         }
+        return redirect(route('home'));
         
     }
 
@@ -84,14 +88,17 @@ class RolesController extends Controller
      */
     public function show(Role $role)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
-                
-           
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
+                    
+               
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
     }
 
     /**
@@ -99,16 +106,19 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
-                $role = Role::where('id',$id)->first();
-                // dd($role->name);
-                return view('admin.edit_role',['role'=>$role]);
-           
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
+                    $role = Role::where('id',$id)->first();
+                    // dd($role->name);
+                    return view('admin.edit_role',['role'=>$role]);
+               
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
        
     }
 
@@ -117,30 +127,33 @@ class RolesController extends Controller
      */
     public function update(Request $request,$id)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
-                $role = Role::where('id',$id)->first();
-        $role->name = $request->role_name;
-        $role->slug = $request->role_slug;
-        $role->save();
-        $role->permissions()->delete();
-        $role->permissions()->detach();
-        $listOfPermissions = explode(',', $request->roles_permissions);//create array from separated/coma permissions
-        
-        foreach ($listOfPermissions as $permission) {
-            $permissions = new Permission();
-            $permissions->name = $permission;
-            $permissions->slug = strtolower(str_replace(" ", "-", $permission));
-            $permissions->save();
-            $role->permissions()->attach($permissions->id);
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Setting" || trim($userPermission->name) == "setting"){
+                    $role = Role::where('id',$id)->first();
+            $role->name = $request->role_name;
             $role->save();
-        }    
-        return redirect(route('userrole'));
-           
+            $role->permissions()->delete();
+            $role->permissions()->detach();
+            $listOfPermissions = explode(',', $request->roles_permissions);//create array from separated/coma permissions
+            
+            foreach ($listOfPermissions as $permission) {
+                $permissions = new Permission();
+                $permissions->name = $permission;
+                // $permissions->slug = strtolower(str_replace(" ", "-", $permission));
+                $permissions->save();
+                $role->permissions()->attach($permissions->id);
+                $role->save();
+            }    
+            return redirect(route('userrole'));
+               
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+       
     }
 
     /**
@@ -148,7 +161,10 @@ class RolesController extends Controller
      */
     public function destroy(Role $role)
     {
-        //
+        if(isset(Auth::user()->username)){
+          
+        }
+        return redirect(route('home'));
     }
 }
