+++ app/Http/Controllers/SliderController.php
@@ -13,8 +13,11 @@ class SliderController extends Controller
      */
     public function index()
     { 
+        if(isset(Auth::user()->username)){
         $sliders = Slider::All();
-       return view('admin.slider',['sliders'=>$sliders]);
+        return view('admin.slider',['sliders'=>$sliders]);
+    }
+    return redirect(route('home'));
     }
 
     /**
@@ -30,38 +33,42 @@ class SliderController extends Controller
      */
     public function store(Request $request)
     {
-        // dd($request->status);
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Create" || trim($userPermission->name) == "create"){
-                $request->validate([
-                    'heading'=>'required',
-                    'content'=>'required',
-                    'image'=>'required'
-                ]);
-                $image = $request->file('image');
-                $imagename = $image->getClientOriginalName();
-                $request->image->move(public_path('image'),$imagename);
-
-                $slide = new Slider();
-                $slide->heading = $request->heading;
-                $slide->content = $request->content;
-                $slide->image = $imagename;
-               
-                $sliders = Slider::All();
-                foreach($sliders as $slider){
-                    if($slider->status == 1 && $request->status != null){
-                       $slide->status = 0;
-                    }else{
-                        $slide->status = 1;
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Create" || trim($userPermission->name) == "create"){
+                    $request->validate([
+                        'heading'=>'required',
+                        'content'=>'required',
+                        'image'=>'required'
+                    ]);
+                    $image = $request->file('image');
+                    $imagename = $image->getClientOriginalName();
+                    $request->image->move(public_path('image'),$imagename);
+    
+                    $slide = new Slider();
+                    $slide->heading = $request->heading;
+                    $slide->content = $request->content;
+                    $slide->image = $imagename;
+                   
+                    $sliders = Slider::All();
+                    foreach($sliders as $slider){
+                        if($slider->status == 1 && $request->status != null || $request->status == null){
+                           $slide->status = 0;
+                        }else{
+                            $slide->status = 1;
+                        }
                     }
+                    // dd('nice');
+                    $slide->save();
+                    return redirect(route('slider'));
                 }
-                // dd('nice');
-                $slide->save();
-                return redirect(route('slider'));
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+        // dd($request->status);
+       
     }
 
     /**
@@ -77,15 +84,17 @@ class SliderController extends Controller
      */
     public function edit(Slider $slider,$id)
     {
-        // dd('hello');
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Create" || trim($userPermission->name) == "create"){
-                $slider = Slider::where('id',$id)->first();
-                return view('admin.edit_slider',['slider'=>$slider]);
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Create" || trim($userPermission->name) == "create"){
+                    $slider = Slider::where('id',$id)->first();
+                    return view('admin.edit_slider',['slider'=>$slider]);
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
        
     }
 
@@ -94,37 +103,41 @@ class SliderController extends Controller
      */
     public function update(Request $request, Slider $slider,$id)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
-                $request->validate([
-                    'heading'=>'required',
-                    'content'=>'required',
-                ]);
-
-                $slide = Slider::where('id',$id)->first();
-                $slide->heading = $request->heading;
-                $slide->content = $request->content;
-                if($request->image != null){
-                    $image = $request->file('image');
-                    $imagename = $image->getClientOriginalName();
-                    $request->image->move(public_path('image'),$imagename);
-                    $slide->image = $imagename;
-                }
-                $sliders = Slider::All();
-                foreach($sliders as $slider){
-                    if($slider->status == 1 && $request->status != null){
-                       $slide->status = 0;
-                    }else{
-                        $slide->status = 1;
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Edit" || trim($userPermission->name) == "edit"){
+                    $request->validate([
+                        'heading'=>'required',
+                        'content'=>'required',
+                    ]);
+    
+                    $slide = Slider::where('id',$id)->first();
+                    $slide->heading = $request->heading;
+                    $slide->content = $request->content;
+                    if($request->image != null){
+                        $image = $request->file('image');
+                        $imagename = $image->getClientOriginalName();
+                        $request->image->move(public_path('image'),$imagename);
+                        $slide->image = $imagename;
                     }
+                    $sliders = Slider::All();
+                    foreach($sliders as $slider){
+                        if($slider->status == 1 && $request->status != null){
+                           $slide->status = 0;
+                        }else{
+                            $slide->status = 1;
+                        }
+                    }
+                    // dd('nice');
+                    $slide->save();
+                    return redirect(route('slider'));
                 }
-                // dd('nice');
-                $slide->save();
-                return redirect(route('slider'));
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
+        
     }
 
     /**
@@ -132,14 +145,17 @@ class SliderController extends Controller
      */
     public function destroy(Slider $slider,$id)
     {
-        $userPermissions = Auth::user()->permissions;
-        foreach($userPermissions as $userPermission){
-            if(trim($userPermission->name) == "Delete" || trim($userPermission->name) == "delete"){
-                $slider = Slider::where('id',$id)->first();
-                $slider->delete();
-                return redirect(route('slider'));
+        if(isset(Auth::user()->username)){
+            $userPermissions = Auth::user()->permissions;
+            foreach($userPermissions as $userPermission){
+                if(trim($userPermission->name) == "Delete" || trim($userPermission->name) == "delete"){
+                    $slider = Slider::where('id',$id)->first();
+                    $slider->delete();
+                    return redirect(route('slider'));
+                }
             }
+            return redirect(route('dashboard'));
         }
-        return redirect(route('dashboard'));
+        return redirect(route('home'));
     }
 }
